// src/enums/position.ts
export enum Position {
    JUNIOR_DEVELOPER = 'JUNIOR_DEVELOPER',
    SENIOR_DEVELOPER = 'SENIOR_DEVELOPER',
    TEAM_LEAD = 'TEAM_LEAD',
    MANAGER = 'MANAGER',
    DIRECTOR = 'DIRECTOR',
    CEO = 'CEO',
    HR_SPECIALIST = 'HR_SPECIALIST',
    ACCOUNTANT = 'ACCOUNTANT',
    MARKETING_SPECIALIST = 'MARKETING_SPECIALIST',
    SALES_REPRESENTATIVE = 'SALES_REPRESENTATIVE',
} 